+++ plantdescriptionengine/src/main/java/eu/arrowhead/core/plantdescriptionengine/services/pde_monitor/dto/PlantDescriptionEntry.java
@@ -22,25 +22,25 @@ public interface PlantDescriptionEntry {
     final static Comparator<PlantDescriptionEntry> ID_COMPARATOR = new Comparator<>() {
         @Override
         public int compare(PlantDescriptionEntry e1, PlantDescriptionEntry e2) {
-            return e1.id() - e2.id();
+            return (int)(e1.id() - e2.id());
         }
     };
 
     final static Comparator<PlantDescriptionEntry> CREATED_AT_COMPARATOR = new Comparator<>() {
         @Override
         public int compare(PlantDescriptionEntry e1, PlantDescriptionEntry e2) {
-            return e1.createdAt().compareTo(e2.createdAt());
+            return (int)(e1.createdAt().compareTo(e2.createdAt()));
         }
     };
 
     final static Comparator<PlantDescriptionEntry> UPDATED_AT_COMPARATOR = new Comparator<>() {
         @Override
         public int compare(PlantDescriptionEntry e1, PlantDescriptionEntry e2) {
-            return e1.updatedAt().compareTo(e2.updatedAt());
+            return (int)(e1.updatedAt().compareTo(e2.updatedAt()));
         }
     };
 
-    int id();
+    float id();
     String plantDescription();
     boolean active();
     List<Float> include();
